requirements.txt
langchain
langgraph
transformers
sentence-transformers
python-dotenv
pandas
plotly
nbformat
bs4
faiss-cpu
lxml
pypdf